+++ packages/calcite-components/src/components/action-menu/action-menu.e2e.ts
@@ -1,5 +1,5 @@
-import { newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
-import { describe, expect, it } from "vitest";
+import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
+import { beforeEach, describe, expect, it } from "vitest";
 import { html } from "../../../support/formatting";
 import {
   accessible,
@@ -200,6 +200,53 @@ describe("calcite-action-menu", () => {
 
     const focusTargetSelector = `#triggerAction`;
     await isElementFocused(page, focusTargetSelector);
+  });
+
+  describe("adding/removing from DOM", () => {
+    let page: E2EPage;
+
+    beforeEach(async (): Promise<void> => {
+      page = await newE2EPage();
+      await skipAnimations(page);
+    });
+
+    async function testToggle(triggerSelector: string): Promise<void> {
+      await page.evaluate(() => {
+        const actionMenu = document.querySelector("calcite-action-menu");
+        actionMenu.remove();
+        document.body.append(actionMenu);
+      });
+      await page.waitForChanges();
+
+      const trigger = await page.find(triggerSelector);
+      await trigger.click();
+
+      const actionMenu = await page.find("calcite-action-menu");
+      expect(await actionMenu.getProperty("open")).toBe(true);
+    }
+
+    it("should toggle with default trigger", async () => {
+      await page.setContent(html`
+        <calcite-action-menu>
+          <calcite-action text="Add" icon="plus" text-enabled></calcite-action>
+          <calcite-action text="Remove" icon="minus" text-enabled></calcite-action>
+          <calcite-action text="Banana" icon="banana" text-enabled></calcite-action>
+        </calcite-action-menu>
+      `);
+      await testToggle(`calcite-action-menu >>> .${CSS.defaultTrigger}`);
+    });
+
+    it("should toggle with slotted trigger", async () => {
+      await page.setContent(html`
+        <calcite-action-menu>
+          <calcite-action id="trigger" slot="${SLOTS.trigger}" text="Toggle" icon="toggle"></calcite-action>
+          <calcite-action text="Add" icon="plus" text-enabled></calcite-action>
+          <calcite-action text="Remove" icon="minus" text-enabled></calcite-action>
+          <calcite-action text="Banana" icon="banana" text-enabled></calcite-action>
+        </calcite-action-menu>
+      `);
+      await testToggle("#trigger");
+    });
   });
 
   it("should honor scale of expand icon", async () => {