+++ packages/calcite-components/src/components/autocomplete/autocomplete.tsx
@@ -8,9 +8,10 @@ import {
   state,
   JsxNode,
   stringOrBoolean,
+  LuminaJsx,
 } from "@arcgis/lumina";
 import { useWatchAttributes } from "@arcgis/components-controllers";
-import { debounce } from "lodash-es";
+import { debounce, escapeRegExp } from "lodash-es";
 import {
   FlipPlacement,
   FloatingCSS,
@@ -102,6 +103,8 @@ export class Autocomplete
 
   defaultValue: Autocomplete["value"];
 
+  defaultInputValue: Autocomplete["inputValue"];
+
   floatingEl: HTMLDivElement;
 
   floatingLayout?: FloatingLayout;
@@ -126,6 +129,8 @@ export class Autocomplete
   referenceEl: Input["el"];
 
   transitionEl: HTMLDivElement;
+
+  private inputValueMatchPattern: RegExp;
 
   // #endregion
 
@@ -164,9 +169,9 @@ export class Autocomplete
    * Specifies the type of content to autocomplete, for use in forms.
    * Read the native attribute's documentation on MDN for more info.
    *
-   * @mdn [step](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete)
+   * @mdn [autocomplete](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete)
    */
-  @property() autocomplete: string;
+  @property() autocomplete: AutoFill;
 
   /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
   @property({ reflect: true }) disabled = false;
@@ -338,6 +343,23 @@ export class Autocomplete
   }
 
   /**
+   * Scrolls the component's content to a specified set of coordinates.
+   *
+   * @example
+   * myAutocomplete.scrollContentTo({
+   *   left: 0, // Specifies the number of pixels along the X axis to scroll the window or element.
+   *   top: 0, // Specifies the number of pixels along the Y axis to scroll the window or element
+   *   behavior: "auto" // Specifies whether the scrolling should animate smoothly (smooth), or happen instantly in a single jump (auto, the default value).
+   * });
+   * @param options - allows specific coordinates to be defined.
+   * @returns - promise that resolves once the content is scrolled to.
+   */
+  @method()
+  async scrollContentTo(options?: ScrollToOptions): Promise<void> {
+    this.transitionEl?.scrollTo(options);
+  }
+
+  /**
    * Selects the text of the component's `value`.
    *
    * @returns {Promise<void>}
@@ -396,6 +418,7 @@ export class Autocomplete
     this.mutationObserver?.observe(this.el, { childList: true, subtree: true });
     connectLabel(this);
     connectForm(this);
+    this.defaultInputValue = this.inputValue || "";
 
     this.getAllItemsDebounced();
 
@@ -440,12 +463,27 @@ export class Autocomplete
       this.reposition(true);
     }
 
-    if (changes.has("scale") && (this.hasUpdated || this.scale !== "m")) {
+    let itemsAndGroupsUpdated = false;
+
+    if (changes.has("inputValue") && (this.hasUpdated || this.inputValue)) {
+      this.inputValueMatchPattern =
+        this.inputValue && new RegExp(`(${escapeRegExp(this.inputValue)})`, "i");
       this.updateItems();
       this.updateGroups();
+      itemsAndGroupsUpdated = true;
     }
 
-    if (changes.has("activeIndex") && (this.hasUpdated || this.activeIndex !== -1)) {
+    if (!itemsAndGroupsUpdated && changes.has("scale") && (this.hasUpdated || this.scale !== "m")) {
+      this.updateItems();
+      this.updateGroups();
+      itemsAndGroupsUpdated = true;
+    }
+
+    if (
+      !itemsAndGroupsUpdated &&
+      changes.has("activeIndex") &&
+      (this.hasUpdated || this.activeIndex !== -1)
+    ) {
       this.updateItems();
     }
   }
@@ -456,6 +494,7 @@ export class Autocomplete
 
   loaded(): void {
     afterConnectDefaultValueSet(this, this.value || "");
+    this.defaultInputValue = this.inputValue || "";
     setComponentLoaded(this);
     connectFloatingUI(this);
   }
@@ -516,10 +555,12 @@ export class Autocomplete
     this.open = false;
   }
 
-  private handleInternalAutocompleteItemSelect(event: Event): void {
+  private async handleInternalAutocompleteItemSelect(event: Event): Promise<void> {
     this.value = (event.target as AutocompleteItem["el"]).value;
     event.stopPropagation();
     this.emitChange();
+    await this.setFocus();
+    this.open = false;
   }
 
   private mutationObserver = createObserver("mutation", () => this.getAllItemsDebounced());
@@ -530,6 +571,10 @@ export class Autocomplete
 
   onLabelClick(): void {
     this.setFocus();
+  }
+
+  onFormReset(): void {
+    this.inputValue = this.defaultInputValue;
   }
 
   onBeforeOpen(): void {
@@ -549,7 +594,6 @@ export class Autocomplete
   }
 
   private emitChange(): void {
-    this.open = false;
     this.calciteAutocompleteChange.emit();
   }
 
@@ -581,6 +625,7 @@ export class Autocomplete
 
       item.active = isActive;
       item.scale = this.scale;
+      item.inputValueMatchPattern = this.inputValueMatchPattern;
     });
 
     this.activeDescendant = activeDescendant;
@@ -624,11 +669,6 @@ export class Autocomplete
     this.resizeObserver?.observe(el);
 
     connectFloatingUI(this);
-
-    // TODO: fixme when supported in jsx
-    // https://devtopia.esri.com/WebGIS/arcgis-web-components/issues/2694
-    const enterKeyHint = this.el.getAttribute("enterkeyhint");
-    el.enterKeyHint = enterKeyHint;
   }
 
   private keyDownHandler(event: KeyboardEvent): void {
@@ -654,6 +694,7 @@ export class Autocomplete
         if (open && activeIndex > -1) {
           this.value = enabledItems[activeIndex].value;
           this.emitChange();
+          this.open = false;
           event.preventDefault();
         } else if (!event.defaultPrevented) {
           if (submitForm(this)) {
@@ -665,31 +706,47 @@ export class Autocomplete
         this.open = true;
         this.activeIndex =
           activeIndex !== -1 ? Math.min(activeIndex + 1, enabledItems.length - 1) : 0;
+        this.scrollToActiveItem();
         event.preventDefault();
         break;
       case "ArrowUp":
         this.open = true;
         this.activeIndex =
           activeIndex !== -1 ? Math.max(activeIndex - 1, 0) : enabledItems.length - 1;
+        this.scrollToActiveItem();
         event.preventDefault();
         break;
       case "Home":
         this.open = true;
         this.activeIndex = 0;
+        this.scrollToActiveItem();
         event.preventDefault();
         break;
       case "End":
         this.open = true;
         this.activeIndex = enabledItems.length - 1;
+        this.scrollToActiveItem();
         event.preventDefault();
         break;
     }
+  }
+
+  private scrollToActiveItem(): void {
+    this.enabledItems[this.activeIndex]?.scrollIntoView({ block: "nearest" });
   }
 
   private changeHandler(event: CustomEvent): void {
     event.stopPropagation();
     this.inputValue = (event.target as Input["el"]).value;
     this.calciteAutocompleteTextChange.emit();
+  }
+
+  private inputClickHandler(event: MouseEvent): void {
+    if (event.defaultPrevented) {
+      return;
+    }
+
+    this.open = true;
   }
 
   private inputHandler(event: CustomEvent): void {
@@ -714,16 +771,9 @@ export class Autocomplete
   override render(): JsxNode {
     const { disabled, listId, inputId, isOpen } = this;
 
-    const autofocus = this.el.autofocus || this.el.hasAttribute("autofocus") ? true : null;
-    const inputMode = this.el.getAttribute("inputmode") as
-      | "none"
-      | "text"
-      | "search"
-      | "email"
-      | "tel"
-      | "url"
-      | "numeric"
-      | "decimal";
+    const autofocus = this.el.autofocus;
+    const enterKeyHint = this.el.enterKeyHint as LuminaJsx.HTMLElementTags["input"]["enterKeyHint"];
+    const inputMode = this.el.inputMode as LuminaJsx.HTMLElementTags["input"]["inputMode"];
 
     return (
       <InteractiveContainer disabled={disabled}>
@@ -741,6 +791,7 @@ export class Autocomplete
             class={CSS.input}
             clearable={true}
             disabled={disabled}
+            enterKeyHint={enterKeyHint}
             form={this.form}
             icon={this.getIcon()}
             iconFlipRtl={this.iconFlipRtl}
@@ -752,6 +803,7 @@ export class Autocomplete
             messageOverrides={this.messages}
             minLength={this.minLength}
             name={this.name}
+            onClick={this.inputClickHandler}
             onKeyDown={this.keyDownHandler}
             oncalciteInputChange={this.changeHandler}
             oncalciteInputInput={this.inputHandler}